Feat/basket

+++ Back-end/src/main/java/com/simplon/dvdStore/repositories/ClientRepository.java
@@ -1,6 +1,10 @@
 package com.simplon.dvdStore.repositories;
 
+import com.simplon.dvdStore.domain.ClientRepositoryModelSQL;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepository extends CrudRepository<ClientRepositoryModelSQL, Long> {
+/**
+ * The interface Client repository.
+ */
+public interface ClientRepository extends CrudRepository<ClientRepositoryModelSQL, Integer> {
 }
