+++ Back-end/src/main/java/com/simplon/dvdStore/repositories/DvdRepository.java
@@ -1,6 +1,10 @@
 package com.simplon.dvdStore.repositories;
 
+import com.simplon.dvdStore.domain.DvdRepositoryModelSQL;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DvdRepository extends CrudRepository<DvdRepositoryModelSQL, Long> {
+/**
+ * The interface Dvd repository.
+ */
+public interface DvdRepository extends CrudRepository<DvdRepositoryModelSQL, Integer> {
 }
