+++ Back-end/src/main/java/com/simplon/dvdStore/services/ClientServiceModel.java
@@ -6,10 +6,13 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 
+/**
+ * The type Client service model.
+ */
 @Data
 @NoArgsConstructor
 public class ClientServiceModel {
-    private Long id;
+    private int id;
     private String clientReference;
     private String firstName;
     private String lastName;
